Refactor package layout and move test fixtures to testdata
