+++ fcrypt.go
@@ -1,988 +1,13 @@
 package fcrypt
 
-import (
-	"context"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/rand"
-	"crypto/sha256"
-	"crypto/sha512"
-	"encoding/binary"
-	"encoding/hex"
-	"errors"
-	"fmt"
-	"hash"
-	"io"
-	"os"
-
-	"golang.org/x/crypto/blake2b"
-	"golang.org/x/crypto/scrypt"
-	"golang.org/x/crypto/sha3"
-)
-
-// Key interface to provide methods for handling encryption keys
-type Key interface {
-	// Encryption key version
-	Version() string
-	// Salt used for key derivation
-	Salt() []byte
-	// Encryption algorithm
-	Algo() string
-	// Encryption key as byte slice
-	KeyBytes() []byte
-}
-
-// FcryptKey struct implements the Key interface
-type FcryptKey struct {
-	version string
-	salt    []byte
-	algo    string
-	key     []byte
-}
-
-// SetVersion sets the version of the FcryptKey.
-func (k *FcryptKey) SetVersion(v string) {
-	k.version = v
-}
-
-// Version returns the version of the FcryptKey.
-func (k *FcryptKey) Version() string {
-	return k.version
-}
-
-// SetSalt sets the salt value for the FcryptKey.
-// The salt is used as an additional input to the key derivation function,
-// making it harder to perform precomputed dictionary attacks.
-func (k *FcryptKey) SetSalt(s []byte) {
-	k.salt = s
-}
-
-// Salt returns the salt value associated with the FcryptKey.
-func (k *FcryptKey) Salt() []byte {
-	return k.salt
-}
-
-// SetAlgo sets the encryption algorithm for the FcryptKey.
-// The algorithm should be a string representing the desired encryption algorithm.
-func (k *FcryptKey) SetAlgo(a string) {
-	k.algo = a
-}
-
-// Algo returns the algorithm used by the FcryptKey.
-func (k *FcryptKey) Algo() string {
-	return k.algo
-}
-
-// SetKeyBytes sets the key bytes for the FcryptKey instance.
-// The key parameter is a byte slice containing the key bytes.
-func (k *FcryptKey) SetKeyBytes(key []byte) {
-	k.key = key
-}
-
-// KeyBytes returns the key bytes of the FcryptKey.
-func (k *FcryptKey) KeyBytes() []byte {
-	return k.key
-}
-
-// SetAll sets the values of the FcryptKey struct.
-// It takes in the version string, salt byte slice, algorithm string, and key byte slice as parameters.
-func (k *FcryptKey) SetAll(v string, s []byte, a string, key []byte) {
-	k.version = v
-	k.salt = s
-	k.algo = a
-	k.key = key
-}
-
-// NewFcryptKey creates a new FcryptKey with the specified version, salt, algorithm, and key.
-func NewFcryptKey(version string, salt []byte, algo string, key []byte) *FcryptKey {
-	return &FcryptKey{
-		version: version,
-		salt:    salt,
-		algo:    algo,
-		key:     key,
-	}
-}
-
-// Constants and errors
-const (
-	// MinKeyLength is the minimum length of the encryption key in bytes.
-	MinKeyLength = 16
-	// DefaultKeyLength is the default length of the encryption key in bytes.
-	DefaultKeyLength = 32
-	// ScryptN is the CPU/memory cost parameter for scrypt.
-	ScryptN = 32768
-	// ScryptR is the block size parameter for scrypt.
-	ScryptR = 8
-	// ScryptP is the parallelization parameter for scrypt.
-	ScryptP = 1
-	// MinNonceSize is the minimum size of the nonce in bytes.
-	MinNonceSize = 12
-	// GCMNonceSize is the size of the nonce used in GCM mode.
-	GCMNonceSize = 12
-)
-
-// Error variables
-var (
-	ErrCiphertextTooShort   = errors.New("ciphertext too short")
-	ErrKeyLengthTooShort    = errors.New("key length too short")
-	ErrFailedToCreateCipher = errors.New("failed to create new cipher")
-	ErrFailedToCreateGCM    = errors.New("failed to create new GCM")
-	ErrFailedToCreateFile   = errors.New("failed to create file")
-	ErrFailedToReadData     = errors.New("failed to read data")
-)
-
-// GenerateSalt generates a random salt of the specified length.
-// It uses the crypto/rand package to generate cryptographically secure random bytes.
-// The length parameter specifies the number of bytes to generate.
-// It returns the generated salt as a byte slice and any error encountered during the generation process.
-func GenerateSalt(length int) ([]byte, error) {
-	salt := make([]byte, length)
-	if _, err := rand.Read(salt); err != nil {
-		return nil, err
-	}
-	return salt, nil
-}
-
-// GenerateKey generates a key using the provided passphrase, salt, and key length.
-// It uses the scrypt key derivation function to derive the key from the passphrase and salt.
-// The key length specifies the desired length of the generated key in bytes.
-// Returns the generated key as a byte slice and any error encountered during the key generation process.
-func GenerateKey(passphrase string, salt []byte, keyLength int) ([]byte, error) {
-	// check if the key length is valid
-	if keyLength <= MinKeyLength {
-		return nil, ErrKeyLengthTooShort
-	}
-	keyBytes, err := scrypt.Key([]byte(passphrase), salt, ScryptN, ScryptR, ScryptP, keyLength)
-	if err != nil {
-		return nil, err
-	}
-	return keyBytes, nil
-}
-
-// GenerateSaltAndKey generates the salt and key using the provided passphrase and key length.
-// It first generates the salt of the specified length and then derives the key using that salt.
-func GenerateSaltAndKey(passphrase string, saltLength int, keyLength int) ([]byte, []byte, error) {
-	salt, err := GenerateSalt(saltLength)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	key, err := GenerateKey(passphrase, salt, keyLength)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return salt, key, nil
-}
-
-// GenerateGCM generates a Galois/Counter Mode (GCM) cipher.AEAD and cipher.Block using the provided key.
-// It returns the generated gcm, block, and any error encountered during the process.
-func GenerateGCM(key []byte) (gcm cipher.AEAD, block cipher.Block, err error) {
-	block, err = aes.NewCipher(key)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-
-	gcm, err = cipher.NewGCM(block)
-	if err != nil {
-		return nil, nil, fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-	return
-}
-
-// GenerateGCMWithNonce generates a Galois/Counter Mode (GCM) cipher with a random nonce equal to the GCM nonce size.
-// It takes a key as input and returns the GCM cipher, the underlying block cipher,
-// the generated nonce, and any error that occurred during the process.
-func GenerateGCMWithNonce(key []byte) (gcm cipher.AEAD, block cipher.Block, nonce []byte, err error) {
-	gcm, block, err = GenerateGCM(key)
-	if err != nil {
-		return
-	}
-
-	nonce = make([]byte, gcm.NonceSize())
-	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
-		return nil, nil, nil, fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-	}
-	return
-}
-
-// Encrypt encrypts the given data using the provided key and returns the encrypted result.
-// It uses the GCM mode of operation for encryption.
-// The nonce is randomly generated and prepended to the encrypted data.
-func Encrypt(data []byte, key []byte) ([]byte, error) {
-	gcm, _, err := GenerateGCM(key)
-	if err != nil {
-		return nil, err
-	}
-
-	nonce := make([]byte, gcm.NonceSize())
-	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
-		return nil, err
-	}
-
-	return gcm.Seal(nonce, nonce, data, nil), nil
-}
-
-// Decrypt decrypts the given ciphertext using the provided key.
-// It returns the plaintext or an error if decryption fails.
-func Decrypt(data []byte, key []byte) ([]byte, error) {
-	gcm, _, err := GenerateGCM(key)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(data) < gcm.NonceSize() {
-		return nil, ErrCiphertextTooShort
-	}
-
-	nonce, ciphertext := data[:gcm.NonceSize()], data[gcm.NonceSize():]
-	return gcm.Open(nil, nonce, ciphertext, nil)
-}
-
-// ReEncrypt re-encrypts the given data using the oldKey and then encrypts it again using the newKey.
-// It returns the re-encrypted data or an error if the encryption process fails.
-func ReEncrypt(data []byte, oldKey []byte, newKey []byte) ([]byte, error) {
-	decryptedData, err := Decrypt(data, oldKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return Encrypt(decryptedData, newKey)
-}
-
-// EncryptChunk encrypts the given plaintext using the provided block cipher and nonce.
-// It returns the ciphertext and an error, if any.
-func EncryptChunk(block cipher.Block, plaintext []byte, nonce []byte) ([]byte, error) {
-	aesgcm, err := cipher.NewGCM(block)
-	if err != nil {
-		return nil, fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	return aesgcm.Seal(nil, nonce, plaintext, nil), nil
-}
-
-// EncryptFileToFile encrypts the data from the given reader using the provided key and writes it to the specified file.
-// The encryption is done in chunks of the specified size. It uses AES encryption with GCM mode.
-// The function returns an error if any operation fails.
-func EncryptFileToFile(data io.Reader, key []byte, chunkSize int, filePath string) error {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-
-	aead, err := cipher.NewGCM(block)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	file, err := os.Create(filePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer file.Close()
-
-	chunk := make([]byte, chunkSize)
-	nonce := make([]byte, aead.NonceSize())
-	lenBuf := make([]byte, 4)
-
-	for {
-		n, readErr := data.Read(chunk)
-		if readErr == io.EOF {
-			break
-		}
-		if readErr != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, readErr)
-		}
-
-		if _, err := io.ReadFull(rand.Reader, nonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		ciphertext := aead.Seal(nil, nonce, chunk[:n], nil)
-		if len(ciphertext) > int(^uint32(0)) {
-			return fmt.Errorf("ciphertext chunk too large: %d", len(ciphertext))
-		}
-
-		binary.BigEndian.PutUint32(lenBuf, uint32(len(ciphertext)))
-
-		// Record format: nonce || uint32(ciphertextLen) || ciphertext
-		if _, err := file.Write(nonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := file.Write(lenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := file.Write(ciphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-	}
-
-	return nil
-}
-
-// DecryptFileToFile decrypts the contents of an encrypted file and writes the decrypted data to a new file.
-// It takes the path of the encrypted file, the path of the decrypted file, the encryption key, and the chunk size as parameters.
-// The function reads the encrypted file in chunks, decrypts each chunk using AES-GCM encryption, and writes the decrypted data to the new file.
-// It returns an error if any operation fails.
-func DecryptFileToFile(encryptedFilePath, decryptedFilePath string, key []byte, chunkSize int) error {
-	encryptedFile, err := os.Open(encryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer encryptedFile.Close()
-
-	decryptedFile, err := os.Create(decryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer decryptedFile.Close()
-
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-
-	aead, err := cipher.NewGCM(block)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	nonce := make([]byte, aead.NonceSize())
-	lenBuf := make([]byte, 4)
-
-	for {
-		// Read nonce
-		if _, err := io.ReadFull(encryptedFile, nonce); err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		// Read ciphertext length
-		if _, err := io.ReadFull(encryptedFile, lenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		ctLen := binary.BigEndian.Uint32(lenBuf)
-		if ctLen == 0 {
-			return fmt.Errorf("invalid ciphertext length: 0")
-		}
-
-		ciphertext := make([]byte, int(ctLen))
-		if _, err := io.ReadFull(encryptedFile, ciphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		plaintext, err := aead.Open(nil, nonce, ciphertext, nil)
-		if err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-		}
-
-		if _, err := decryptedFile.Write(plaintext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-	}
-
-	_ = chunkSize // kept for API compatibility; chunking is implicit in the encrypted record stream
-	return nil
-}
-
-// ReEncryptFileToFile re-encrypts the contents of an encrypted file using a new encryption key
-// and writes the decrypted contents to a new file.
+// Package fcrypt provides ergonomic helpers for encrypting/decrypting bytes and files,
+// key derivation/rotation, and hashing utilities.
 //
-// Parameters:
-// - encryptedFilePath: The path to the encrypted file.
-// - decryptedFilePath: The path to the new file where the decrypted contents will be written.
-// - oldKey: The old encryption key used to encrypt the file.
-// - newKey: The new encryption key to be used for re-encryption.
-// - chunkSize: The size of each chunk to be read from the encrypted file.
-//
-// Returns:
-// - An error if any error occurs during the re-encryption process, or nil if the re-encryption is successful.
-//
-// Example usage:
-//
-//	err := ReEncryptFileToFile("/path/to/encrypted/file", "/path/to/decrypted/file", oldKey, newKey, 1024)
-//	if err != nil {
-//		fmt.Println("Error:", err)
-//	}
-func ReEncryptFileToFile(encryptedFilePath, decryptedFilePath string, oldKey []byte, newKey []byte, chunkSize int) error {
-	encryptedFile, err := os.Open(encryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer encryptedFile.Close()
-
-	outFile, err := os.Create(decryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer outFile.Close()
-
-	oldBlock, err := aes.NewCipher(oldKey)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-	oldAEAD, err := cipher.NewGCM(oldBlock)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	newBlock, err := aes.NewCipher(newKey)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-	newAEAD, err := cipher.NewGCM(newBlock)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	nonce := make([]byte, oldAEAD.NonceSize())
-	lenBuf := make([]byte, 4)
-	newNonce := make([]byte, newAEAD.NonceSize())
-	outLenBuf := make([]byte, 4)
-
-	for {
-		if _, err := io.ReadFull(encryptedFile, nonce); err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		if _, err := io.ReadFull(encryptedFile, lenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		ctLen := binary.BigEndian.Uint32(lenBuf)
-		if ctLen == 0 {
-			return fmt.Errorf("invalid ciphertext length: 0")
-		}
-
-		ciphertext := make([]byte, int(ctLen))
-		if _, err := io.ReadFull(encryptedFile, ciphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		plaintext, err := oldAEAD.Open(nil, nonce, ciphertext, nil)
-		if err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-		}
-
-		if _, err := io.ReadFull(rand.Reader, newNonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		newCiphertext := newAEAD.Seal(nil, newNonce, plaintext, nil)
-		binary.BigEndian.PutUint32(outLenBuf, uint32(len(newCiphertext)))
-
-		if _, err := outFile.Write(newNonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := outFile.Write(outLenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := outFile.Write(newCiphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-	}
-
-	_ = chunkSize
-	return nil
-}
-
-// StreamEncrypt takes an input data stream and a key, and returns an encrypted data stream along with any error encountered.
-// The function generates a GCM (Galois/Counter Mode) cipher using the provided key, and then generates a random nonce.
-// It uses the GCM cipher in CTR (Counter) mode to create a cipher stream reader, which encrypts the input data stream.
-// The encrypted data stream is returned along with a possible error.
-func StreamEncrypt(data io.Reader, key []byte) (io.Reader, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-
-	nonce := make([]byte, block.BlockSize())
-	if _, err := io.ReadFull(rand.Reader, nonce); err != nil {
-		return nil, fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-	}
-
-	pr, pw := io.Pipe()
-	go func() {
-		defer pw.Close()
-		// Write the nonce to the pipe writer
-		if _, err := pw.Write(nonce); err != nil {
-			pw.CloseWithError(err)
-			return
-		}
-		stream := cipher.NewCTR(block, nonce)
-		writer := &cipher.StreamWriter{S: stream, W: pw}
-		if _, err := io.Copy(writer, data); err != nil {
-			pw.CloseWithError(err)
-		}
-	}()
-	return pr, nil
-}
-
-// StreamDecrypt decrypts the data from the given io.Reader using the provided key.
-// It returns an io.Reader that can be used to read the decrypted data, along with any error encountered.
-// The decryption is performed using the AES-GCM mode of operation.
-// The key parameter is the secret key used for decryption.
-// The data parameter is the encrypted data that needs to be decrypted.
-// The returned io.Reader can be used to read the decrypted data.
-// If an error occurs during decryption, it is returned along with a nil io.Reader.
-func StreamDecrypt(data io.Reader, key []byte) (io.Reader, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-
-	nonce := make([]byte, block.BlockSize())
-	if _, err := io.ReadFull(data, nonce); err != nil {
-		return nil, fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-	}
-
-	pr, pw := io.Pipe()
-	go func() {
-		defer pw.Close()
-		stream := cipher.NewCTR(block, nonce)
-		reader := &cipher.StreamReader{S: stream, R: data}
-		if _, err := io.Copy(pw, reader); err != nil {
-			pw.CloseWithError(err)
-		}
-	}()
-	return pr, nil
-}
-
-// StreamReEncrypt re-encrypts the data from the given reader using the oldKey and then encrypts it again using the newKey.
-// It returns an io.Reader containing the re-encrypted data.
-// If any error occurs during the decryption or encryption process, it returns nil and the error.
-func StreamReEncrypt(data io.Reader, oldKey []byte, newKey []byte) (io.Reader, error) {
-	decryptedStream, err := StreamDecrypt(data, oldKey)
-	if err != nil {
-		return nil, err
-	}
-
-	return StreamEncrypt(decryptedStream, newKey)
-}
-
-// Hashing functions
-
-// Hash calculates the hash of the given io.Reader using the provided hash.Hash.
-func Hash(reader io.Reader, hasher hash.Hash) ([]byte, error) {
-	if _, err := io.Copy(hasher, reader); err != nil {
-		return nil, err
-	}
-	return hasher.Sum(nil), nil
-}
-
-// HashBytes calculates the hash of the given byte slice using the provided hash.Hash.
-func HashBytes(data []byte, hasher hash.Hash) []byte {
-	hasher.Write(data)
-	return hasher.Sum(nil)
-}
-
-// HashBytesToString calculates the hash of the given byte slice using the provided hash.Hash.
-func HashBytesToString(data []byte, hasher hash.Hash) string {
-	return hex.EncodeToString(HashBytes(data, hasher))
-}
-
-// HashString calculates the hash of the given string using the provided hash.Hash.
-func HashString(data string, hasher hash.Hash) []byte {
-	hasher.Write([]byte(data))
-	return hasher.Sum(nil)
-}
-
-// HashStringToString calculates the hash of the given string using the provided hash.Hash.
-func HashStringToString(data string, hasher hash.Hash) string {
-	return hex.EncodeToString(HashString(data, hasher))
-}
-
-// HashFile calculates the hash of the given file using the provided hash.Hash.
-func HashFile(file *os.File, hasher hash.Hash) ([]byte, error) {
-	if _, err := io.Copy(hasher, file); err != nil {
-		return nil, err
-	}
-	return hasher.Sum(nil), nil
-}
-
-// Specific SHA-256 functions
-
-// HashBytesSHA256 calculates the SHA-256 hash of the given data.
-func HashBytesSHA256(data []byte) []byte {
-	return HashBytes(data, sha256.New())
-}
-
-// HashBytesToStringSHA256 converts a byte slice to a string representation of its SHA-256 hash.
-func HashBytesToStringSHA256(data []byte) string {
-	return hex.EncodeToString(HashBytes(data, sha256.New()))
-}
-
-// HashStringSHA256 hashes the given string using SHA-256 algorithm and returns the resulting hash as a byte slice.
-func HashStringSHA256(data string) []byte {
-	return HashString(data, sha256.New())
-}
-
-// HashStringToStringSHA256 hashes a string using SHA-256 algorithm and returns the hashed value as a string.
-func HashStringToStringSHA256(data string) string {
-	return HashStringToString(data, sha256.New())
-}
-
-// HashFileSHA256 calculates the SHA-256 hash of the given file.
-func HashFileSHA256(file *os.File) ([]byte, error) {
-	return HashFile(file, sha256.New())
-}
-
-// Specific SHA-512 functions
-
-// HashBytesSHA512 calculates the SHA-512 hash of the given data.
-func HashBytesSHA512(data []byte) []byte {
-	return HashBytes(data, sha512.New())
-}
-
-// HashBytesToStringSHA512 converts a byte slice to a string representation of its SHA-512 hash.
-func HashBytesToStringSHA512(data []byte) string {
-	return hex.EncodeToString(HashBytes(data, sha512.New()))
-}
-
-// HashStringSHA512 hashes the given string using SHA-512 algorithm and returns the resulting hash as a byte slice.
-func HashStringSHA512(data string) []byte {
-	return HashString(data, sha512.New())
-}
-
-// HashStringToStringSHA512 hashes a string using SHA-512 algorithm and returns the hashed value as a string.
-func HashStringToStringSHA512(data string) string {
-	return HashStringToString(data, sha512.New())
-}
-
-// HashFileSHA512 calculates the SHA-512 hash of the given file.
-func HashFileSHA512(file *os.File) ([]byte, error) {
-	return HashFile(file, sha512.New())
-}
-
-// Specific SHA3-256 functions
-
-// HashBytesSHA3 calculates the SHA3-256 hash of the given data.
-// It uses the HashBytes function with a new SHA3-256 hash instance.
-func HashBytesSHA3(data []byte) []byte {
-	return HashBytes(data, sha3.New256())
-}
-
-// HashBytesToStringSHA3 converts a byte slice to a string representation of its SHA3 hash.
-// It takes a byte slice `data` as input and returns the hexadecimal string representation of the SHA3 hash.
-func HashBytesToStringSHA3(data []byte) string {
-	return hex.EncodeToString(HashBytes(data, sha3.New256()))
-}
-
-// HashStringSHA3 hashes the given string using SHA3-256 algorithm and returns the resulting hash as a byte slice.
-func HashStringSHA3(data string) []byte {
-	return HashString(data, sha3.New256())
-}
-
-// HashStringToStringSHA3 hashes a string using SHA3-256 algorithm and returns the hashed value as a string.
-func HashStringToStringSHA3(data string) string {
-	return HashStringToString(data, sha3.New256())
-}
-
-// HashFileSHA3 calculates the SHA3-256 hash of the given file.
-// It takes a *os.File as input and returns the hash as a byte slice.
-// If an error occurs during the hashing process, it is returned as the second value.
-func HashFileSHA3(file *os.File) ([]byte, error) {
-	return HashFile(file, sha3.New256())
-}
-
-// HashWithBlake2b512 hashes the contents of the provided io.Reader using BLAKE2b-512.
-// It returns the computed hash as a byte slice.
-func HashWithBlake2b512(reader io.Reader, key []byte) ([]byte, error) {
-	hasher, err := blake2b.New512(key)
-	if err != nil {
-		return nil, err
-	}
-	return Hash(reader, hasher)
-}
-
-// HashWithBlake2b512NoKey calculates the Blake2b-512 hash of the data read from the given reader,
-// without using any key. It internally calls the HashWithBlake2b512 function with a nil key.
-// It returns the hash as a byte slice and any error encountered during the hashing process.
-func HashWithBlake2b512NoKey(reader io.Reader) ([]byte, error) {
-	return HashWithBlake2b512(reader, nil)
-}
-
-// HashWithBlake2b256 hashes the contents of the provided io.Reader using BLAKE2b-256.
-// It returns the computed hash as a byte slice.
-func HashWithBlake2b256(reader io.Reader, key []byte) ([]byte, error) {
-	hasher, err := blake2b.New256(key)
-	if err != nil {
-		return nil, err
-	}
-	return Hash(reader, hasher)
-}
-
-// HashWithBlake2b256NoKey calculates the Blake2b-256 hash of the data read from the given reader,
-// without using any key. It is a convenience function that calls HashWithBlake2b256 with a nil key.
-// The resulting hash is returned as a byte slice.
-// If an error occurs during the hashing process, it is returned along with a nil byte slice.
-func HashWithBlake2b256NoKey(reader io.Reader) ([]byte, error) {
-	return HashWithBlake2b256(reader, nil)
-}
-
-// Context-aware encryption functions
-
-func EncryptWithContext(ctx context.Context, data []byte, key []byte) ([]byte, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, err
-	}
-	// Encrypt is already bounded; we still honor cancellation before/after.
-	out, err := Encrypt(data, key)
-	if err != nil {
-		return nil, err
-	}
-	if err := ctx.Err(); err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func DecryptWithContext(ctx context.Context, data []byte, key []byte) ([]byte, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, err
-	}
-	out, err := Decrypt(data, key)
-	if err != nil {
-		return nil, err
-	}
-	if err := ctx.Err(); err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func ReEncryptWithContext(ctx context.Context, data []byte, oldKey []byte, newKey []byte) ([]byte, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, err
-	}
-	out, err := ReEncrypt(data, oldKey, newKey)
-	if err != nil {
-		return nil, err
-	}
-	if err := ctx.Err(); err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func EncryptFileToFileWithContext(ctx context.Context, data io.Reader, key []byte, chunkSize int, filePath string) error {
-	if err := ctx.Err(); err != nil {
-		return err
-	}
-
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-	aead, err := cipher.NewGCM(block)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	file, err := os.Create(filePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer file.Close()
-
-	chunk := make([]byte, chunkSize)
-	nonce := make([]byte, aead.NonceSize())
-	lenBuf := make([]byte, 4)
-
-	for {
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-
-		n, readErr := data.Read(chunk)
-		if readErr == io.EOF {
-			break
-		}
-		if readErr != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, readErr)
-		}
-
-		if _, err := io.ReadFull(rand.Reader, nonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		ciphertext := aead.Seal(nil, nonce, chunk[:n], nil)
-		if len(ciphertext) > int(^uint32(0)) {
-			return fmt.Errorf("ciphertext chunk too large: %d", len(ciphertext))
-		}
-		binary.BigEndian.PutUint32(lenBuf, uint32(len(ciphertext)))
-
-		if _, err := file.Write(nonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := file.Write(lenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := file.Write(ciphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-	}
-
-	return nil
-}
-
-func DecryptFileToFileWithContext(ctx context.Context, encryptedFilePath, decryptedFilePath string, key []byte, chunkSize int) error {
-	if err := ctx.Err(); err != nil {
-		return err
-	}
-
-	encryptedFile, err := os.Open(encryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer encryptedFile.Close()
-
-	decryptedFile, err := os.Create(decryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer decryptedFile.Close()
-
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-	aead, err := cipher.NewGCM(block)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	nonce := make([]byte, aead.NonceSize())
-	lenBuf := make([]byte, 4)
-
-	for {
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-
-		if _, err := io.ReadFull(encryptedFile, nonce); err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		if _, err := io.ReadFull(encryptedFile, lenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		ctLen := binary.BigEndian.Uint32(lenBuf)
-		if ctLen == 0 {
-			return fmt.Errorf("invalid ciphertext length: 0")
-		}
-
-		ciphertext := make([]byte, int(ctLen))
-		if _, err := io.ReadFull(encryptedFile, ciphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		plaintext, err := aead.Open(nil, nonce, ciphertext, nil)
-		if err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-		}
-
-		if _, err := decryptedFile.Write(plaintext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-	}
-
-	_ = chunkSize
-	return nil
-}
-
-func ReEncryptFileToFileWithContext(ctx context.Context, encryptedFilePath, decryptedFilePath string, oldKey []byte, newKey []byte, chunkSize int) error {
-	if err := ctx.Err(); err != nil {
-		return err
-	}
-
-	encryptedFile, err := os.Open(encryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer encryptedFile.Close()
-
-	outFile, err := os.Create(decryptedFilePath)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-	}
-	defer outFile.Close()
-
-	oldBlock, err := aes.NewCipher(oldKey)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-	oldAEAD, err := cipher.NewGCM(oldBlock)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	newBlock, err := aes.NewCipher(newKey)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateCipher, err)
-	}
-	newAEAD, err := cipher.NewGCM(newBlock)
-	if err != nil {
-		return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-	}
-
-	nonce := make([]byte, oldAEAD.NonceSize())
-	lenBuf := make([]byte, 4)
-	newNonce := make([]byte, newAEAD.NonceSize())
-	outLenBuf := make([]byte, 4)
-
-	for {
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-
-		if _, err := io.ReadFull(encryptedFile, nonce); err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		if _, err := io.ReadFull(encryptedFile, lenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		ctLen := binary.BigEndian.Uint32(lenBuf)
-		if ctLen == 0 {
-			return fmt.Errorf("invalid ciphertext length: 0")
-		}
-
-		ciphertext := make([]byte, int(ctLen))
-		if _, err := io.ReadFull(encryptedFile, ciphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-
-		plaintext, err := oldAEAD.Open(nil, nonce, ciphertext, nil)
-		if err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateGCM, err)
-		}
-
-		if _, err := io.ReadFull(rand.Reader, newNonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToReadData, err)
-		}
-		newCiphertext := newAEAD.Seal(nil, newNonce, plaintext, nil)
-		binary.BigEndian.PutUint32(outLenBuf, uint32(len(newCiphertext)))
-
-		if _, err := outFile.Write(newNonce); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := outFile.Write(outLenBuf); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-		if _, err := outFile.Write(newCiphertext); err != nil {
-			return fmt.Errorf("%w: %v", ErrFailedToCreateFile, err)
-		}
-	}
-
-	_ = chunkSize
-	return nil
-}
+// Implementation is split across multiple files:
+// - crypto.go: AES-GCM helpers and byte-slice Encrypt/Decrypt APIs
+// - file.go: chunked file encryption using a framed record stream
+// - stream.go: streaming APIs
+// - kdf.go: key derivation helpers
+// - hash.go: hashing helpers
+// - parser.go / generate.go: key parsing and generation
+// - types.go / constants.go / errors.go: shared types/constants/errors
