+++ testdata/keys/example_ecdsa.pub
